refactor: Lotto 클래스 model 패키지 구성
